pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "secure-auth-wm"
version = "0.1.0"
description = "Secure content authentication with digital signatures + robust watermarking"
authors = [{name = "Your Name", email = "you@example.com"}]
requires-python = ">=3.10"
dependencies = [
  "cryptography==43.0.1",
  "Pillow==10.4.0",
  "numpy==2.0.2",
  "scipy==1.14.1",
  "opencv-python==4.10.0.84",
  "streamlit==1.38.0",
  "click==8.1.7",
]

[tool.setuptools.packages.find]
where = ["src"]
